feat: surpess returning default Tag instance
In Mapper class, an optional argument is added to the
method "findTargetFor" in order to be able to surpress
the situation of returning the default output.
Such behaviour is useful in cases when the requested tag is not
present among existing classes and if the method always ouputs
something, it is not clear whether the requested tag is present
among the classes or not.