fix: make a dialog be filled in once
I used to fill in dialog using "on" event listener that resulted
in multiple firing of the dialog filling-in event: clicking the first
time, the event gets called once, clicking the second - it will get
called twice etc. This is due to the fact that the event listeners are
added to existing one. Therefore the whole stack of the actions is
executed. This is resolved by changing keyword "on" to be "once": the
action gets called only once.