feat: find a representative element from cursor position
Controller class is augmented with a method that allows
to find its representative element starting from the current
cursor position.

This representative element is then passed to the dialog as
an extra data.

Previously, the representative element was passing only if
the dialog menu was calling by means of a context menu (right
button click) while if the dialog is called clicking the toolbar
icon, no representative were passing to the dialog and the dialog
filling-in methods were searching for the representative element to use
its data.